+++ CHANGELOG.md
@@ -6,13 +6,28 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.2.0] - 2018-12-13
+
+### Changed
+- Layout adjustments on the topbar.
+- Header no longer depends on 'OrderFormContext' and no longer has Alert.
+
+### Fixed
+- Fixes support for extra headers, such as telemarketing.
+
+## [2.1.5] - 2018-12-06
+
 ### Added
 - Add `linkUrl` prop to `Header` component that changes the logo link URL.
+
+### Changed
+- Updated topbar layout and scroll behavior.
 
 ## [2.1.4] - 2018-12-05
 
 ### Fixed
-- Fixed topbar padding bottom on mobile
+- Fixed topbar padding bottom on mobile.
 
 ## [2.1.3] - 2018-12-05
 ### Fixed